refactor: move error files to prepprocessor root dir
