--- src/preprocessor/utils/CMakeLists.txt
@@ -1,3 +1,0 @@
-aux_source_directory(. SRC_FILES)
-
-target_sources(preprocessor PRIVATE ${SRC_FILES})
